+++ beacon-chain/rpc/core/proposer_deposits_test.go
@@ -1,4 +1,4 @@
-package validator
+package core
 
 import "testing"
