+++ beacon-chain/rpc/core/proposer_empty_block_test.go
@@ -1,4 +1,4 @@
-package validator
+package core
 
 import (
 	"testing"
