+++ beacon-chain/rpc/core/proposer_sync_aggregate.go
@@ -1,4 +1,4 @@
-package validator
+package core
 
 import (
 	"bytes"
