+++ beacon-chain/rpc/core/proposer_sync_aggregate_test.go
@@ -1,4 +1,4 @@
-package validator
+package core
 
 import (
 	"bytes"
