+++ beacon-chain/rpc/core/proposer_utils_bench_test.go
@@ -1,4 +1,4 @@
-package validator
+package core
 
 import (
 	"fmt"
