+++ beacon-chain/rpc/core/unblinder_test.go
@@ -1,4 +1,4 @@
-package validator
+package core
 
 import (
 	"context"
