+++ beacon-chain/rpc/prysm/validator/validator_performance_test.go
@@ -63,8 +63,8 @@ func TestServer_GetValidatorPerformance(t *testing.T) {
 				HeadFetcher: &mock.ChainService{
 					State: headState,
 				},
-				GenesisTimeFetcher: &mock.ChainService{Genesis: time.Now().Add(time.Duration(-1*offset) * time.Second)},
-				SyncChecker:        &mockSync.Sync{IsSyncing: false},
+				TimeFetcher: &mock.ChainService{Genesis: time.Now().Add(time.Duration(-1*offset) * time.Second)},
+				SyncChecker: &mockSync.Sync{IsSyncing: false},
 			},
 		}
 		want := &ValidatorPerformanceResponse{
@@ -120,8 +120,8 @@ func TestServer_GetValidatorPerformance(t *testing.T) {
 					// 10 epochs into the future.
 					State: headState,
 				},
-				SyncChecker:        &mockSync.Sync{IsSyncing: false},
-				GenesisTimeFetcher: &mock.ChainService{Genesis: time.Now().Add(time.Duration(-1*offset) * time.Second)},
+				SyncChecker: &mockSync.Sync{IsSyncing: false},
+				TimeFetcher: &mock.ChainService{Genesis: time.Now().Add(time.Duration(-1*offset) * time.Second)},
 			},
 		}
 		c := headState.Copy()
@@ -185,8 +185,8 @@ func TestServer_GetValidatorPerformance(t *testing.T) {
 					// 10 epochs into the future.
 					State: headState,
 				},
-				SyncChecker:        &mockSync.Sync{IsSyncing: false},
-				GenesisTimeFetcher: &mock.ChainService{Genesis: time.Now().Add(time.Duration(-1*offset) * time.Second)},
+				SyncChecker: &mockSync.Sync{IsSyncing: false},
+				TimeFetcher: &mock.ChainService{Genesis: time.Now().Add(time.Duration(-1*offset) * time.Second)},
 			},
 		}
 		c := headState.Copy()
@@ -255,8 +255,8 @@ func TestServer_GetValidatorPerformance(t *testing.T) {
 				HeadFetcher: &mock.ChainService{
 					State: headState,
 				},
-				GenesisTimeFetcher: &mock.ChainService{Genesis: time.Now().Add(time.Duration(-1*offset) * time.Second)},
-				SyncChecker:        &mockSync.Sync{IsSyncing: false},
+				TimeFetcher: &mock.ChainService{Genesis: time.Now().Add(time.Duration(-1*offset) * time.Second)},
+				SyncChecker: &mockSync.Sync{IsSyncing: false},
 			},
 		}
 		want := &ValidatorPerformanceResponse{
@@ -317,8 +317,8 @@ func TestServer_GetValidatorPerformance(t *testing.T) {
 				HeadFetcher: &mock.ChainService{
 					State: headState,
 				},
-				GenesisTimeFetcher: &mock.ChainService{Genesis: time.Now().Add(time.Duration(-1*offset) * time.Second)},
-				SyncChecker:        &mockSync.Sync{IsSyncing: false},
+				TimeFetcher: &mock.ChainService{Genesis: time.Now().Add(time.Duration(-1*offset) * time.Second)},
+				SyncChecker: &mockSync.Sync{IsSyncing: false},
 			},
 		}
 		want := &ValidatorPerformanceResponse{
@@ -379,8 +379,8 @@ func TestServer_GetValidatorPerformance(t *testing.T) {
 				HeadFetcher: &mock.ChainService{
 					State: headState,
 				},
-				GenesisTimeFetcher: &mock.ChainService{Genesis: time.Now().Add(time.Duration(-1*offset) * time.Second)},
-				SyncChecker:        &mockSync.Sync{IsSyncing: false},
+				TimeFetcher: &mock.ChainService{Genesis: time.Now().Add(time.Duration(-1*offset) * time.Second)},
+				SyncChecker: &mockSync.Sync{IsSyncing: false},
 			},
 		}
 		want := &ValidatorPerformanceResponse{
