+++ beacon-chain/rpc/service.go
@@ -240,15 +240,35 @@ func (s *Service) Start() {
 	s.cfg.Router.HandleFunc("/eth/v1/beacon/blob_sidecars/{block_id}", blobServer.Blobs).Methods(http.MethodGet)
 
 	coreService := &core.Service{
-		HeadFetcher:        s.cfg.HeadFetcher,
-		GenesisTimeFetcher: s.cfg.GenesisTimeFetcher,
-		SyncChecker:        s.cfg.SyncService,
-		Broadcaster:        s.cfg.Broadcaster,
-		SyncCommitteePool:  s.cfg.SyncCommitteeObjectPool,
-		OperationNotifier:  s.cfg.OperationNotifier,
-		AttestationCache:   cache.NewAttestationCache(),
-		StateGen:           s.cfg.StateGen,
-		P2P:                s.cfg.Broadcaster,
+		HeadFetcher:            s.cfg.HeadFetcher,
+		TimeFetcher:            s.cfg.GenesisTimeFetcher,
+		SyncChecker:            s.cfg.SyncService,
+		Broadcaster:            s.cfg.Broadcaster,
+		SyncCommitteePool:      s.cfg.SyncCommitteeObjectPool,
+		OperationNotifier:      s.cfg.OperationNotifier,
+		AttestationCache:       cache.NewAttestationCache(),
+		StateGen:               s.cfg.StateGen,
+		P2P:                    s.cfg.Broadcaster,
+		ForkchoiceFetcher:      s.cfg.ForkchoiceFetcher,
+		OptimisticModeFetcher:  s.cfg.OptimisticModeFetcher,
+		Eth1BlockFetcher:       s.cfg.ExecutionChainService,
+		Eth1InfoFetcher:        s.cfg.ExecutionChainService,
+		DepositFetcher:         s.cfg.DepositFetcher,
+		ChainStartFetcher:      s.cfg.ChainStartFetcher,
+		PendingDepositsFetcher: s.cfg.PendingDepositFetcher,
+		SlashingsPool:          s.cfg.SlashingsPool,
+		ExitPool:               s.cfg.ExitPool,
+		AttPool:                s.cfg.AttestationsPool,
+		BLSChangesPool:         s.cfg.BLSChangesPool,
+		ProposerSlotIndexCache: s.cfg.ProposerIdsCache,
+		BeaconDB:               s.cfg.BeaconDB,
+		FinalizationFetcher:    s.cfg.FinalizationFetcher,
+		ExecutionEngineCaller:  s.cfg.ExecutionEngineCaller,
+		BlockBuilder:           s.cfg.BlockBuilder,
+		BlockReceiver:          s.cfg.BlockReceiver,
+		BlockNotifier:          s.cfg.BlockNotifier,
+		ForkFetcher:            s.cfg.ForkFetcher,
+		BlockFetcher:           s.cfg.ExecutionChainService,
 	}
 
 	validatorServer := &validatorv1alpha1.Server{
